+++ commands/toolbox_tools/stap 3 - Schematisatie bewerken/predict_calc_points.py
@@ -5,6 +5,7 @@ import logging
 
 from qgis.core import QgsMapLayerRegistry, QgsFeatureRequest
 from PyQt4.QtCore import QPyNullVariant
+from PyQt4 import QtGui
 
 from ThreeDiToolbox.utils.user_messages import messagebar_message
 from ThreeDiToolbox.views.predict_calc_points_dialog import (
@@ -12,7 +13,6 @@ from ThreeDiToolbox.views.predict_calc_points_dialog import (
 from ThreeDiToolbox.commands.base.custom_command import (
     CustomCommandBase)
 from ThreeDiToolbox.utils.predictions import Predictor
-
 from ThreeDiToolbox.utils import constants
 log = logging.getLogger(__name__)
 
@@ -41,26 +41,40 @@ class CustomCommand(CustomCommandBase):
         self.tool_dialog_widget.exec_()  # block execution
 
     def run_it(self, db_set, db_type):
+        """
+        :param db_set: dict of database settings. Expected keywords:
+                'host': '',
+                'port': '',
+                'name': '',
+                'username': '',
+                'password': '',
+                'schema': '',
+                'database': '',
+                'db_path': ,
+        :param db_type: 'spatialite' or 'postgres'
+        """
 
-        pal = Predictor(db_type)
-        uri = pal.get_uri(**db_set)
-        calc_pnts_lyr = pal.get_layer_from_uri(
-            uri, 'v2_calculation_point', 'the_geom')
-        self.connected_pnts_lyr = pal.get_layer_from_uri(
-            uri, 'v2_connected_pnt', 'the_geom')
-        pal.start_sqalchemy_engine(db_set)
+        predictor = Predictor(db_type)
+        uri = predictor.get_uri(**db_set)
+        calc_pnts_lyr = predictor.get_layer_from_uri(
+            uri, constants.TABLE_NAME_CALC_PNT, 'the_geom')
+        self.connected_pnts_lyr = predictor.get_layer_from_uri(
+            uri, constants.TABLE_NAME_CONN_PNT, 'the_geom')
+        predictor.start_sqalchemy_engine(db_set)
+        if not self.fresh_start(predictor):
+            return
         default_epsg_code = 28992
-        epsg_code = pal.get_epsg_code() or default_epsg_code
+        epsg_code = predictor.get_epsg_code() or default_epsg_code
         log.info(
             "[*] Using epsg code {} to build the calc_type_dict".format(
                 epsg_code)
         )
-        pal.build_calc_type_dict(epsg_code=epsg_code)
+        predictor.build_calc_type_dict(epsg_code=epsg_code)
         transform = None
-        # spatialites are in WGS84 so we need a tranformation
+        # spatialites are in WGS84 so we need a transformation
         if db_type == 'spatialite':
             transform = '{epsg_code}:4326'.format(epsg_code=epsg_code)
-        succces, features = pal.predict_points(
+        succces, features = predictor.predict_points(
             output_layer=calc_pnts_lyr, transform=transform)
 
         if succces:
@@ -73,7 +87,7 @@ class CustomCommand(CustomCommandBase):
                   'is empty?'.format(len(features))
             level = 1
         messagebar_message("Finished", msg, level=level, duration=12)
-        cp_succces, cp_features = pal.fill_connected_pnts_table(
+        cp_succces, cp_features = predictor.fill_connected_pnts_table(
             calc_pnts_lyr=calc_pnts_lyr,
             connected_pnts_lyr=self.connected_pnts_lyr)
         if cp_succces:
@@ -86,3 +100,37 @@ class CustomCommand(CustomCommandBase):
             cp_level = 1
         messagebar_message("Finished", cp_msg, level=cp_level, duration=12)
         log.info('Done predicting calcualtion points.\n' + msg)
+
+    def fresh_start(self, predictor):
+        """
+        Check whether we start off fresh or not. That is, if the
+        calculation and connected points have been calculated before
+        the stale data will be removed from the database after
+        the user has confirmed to do so
+
+        :param predictor: utils.predictions.Predictor instance
+
+        :returns True if we start fresh. In this case all database
+            tables are empty. False otherwise
+        """
+        fresh = True
+        are_empty = []
+        table_names = [
+            constants.TABLE_NAME_CALC_PNT,
+            constants.TABLE_NAME_CONN_PNT
+        ]
+        for tn in table_names:
+            are_empty.append(predictor.threedi_db.table_is_empty(tn))
+        if not all(are_empty):
+            fresh = False
+            reply = QtGui.QMessageBox.question(
+                self.iface.mainWindow(), 'Warning!',
+                'Calculation point and connected point tables are not empty! '
+                'Do you want to delete all their contents?',
+                QtGui.QMessageBox.Yes, QtGui.QMessageBox.No
+            )
+            if reply == QtGui.QMessageBox.Yes:
+                predictor.threedi_db.delete_from(constants.TABLE_NAME_CALC_PNT)
+                predictor.threedi_db.delete_from(constants.TABLE_NAME_CONN_PNT)
+                fresh = True
+        return fresh